Deprecated FITC, FLMA, and TRITC
The historical short names of Fluorescein Isothiocynate, Fluorescein Mercuric Acetate, and Tetramethylrhodamine are FITC, FLMA, and TRITC respectively. FITC, FLMA, and TRITC are the reactive chemical derivatives of the final intended dye product. The final dye product of FITC and FLMA is Fluorescein. The final dye product of TRITC is Rhodamine. Therefore, in order to maintain consistency in our nomenclature, the ISAC DSTF have decided to deprecate the historical short names of FITC, FLMA, and TRITC and replace them with the final dye product.